+++ L2J_Mobius_5.0_Salvation/java/com/l2jmobius/gameserver/network/serverpackets/ExAcquireSkillInfo.java
@@ -35,7 +35,7 @@ public class ExAcquireSkillInfo implements IClientOutgoingPacket
 	private final int _id;
 	private final int _level;
 	private final int _dualClassLevel;
-	private final int _spCost;
+	private final long _spCost;
 	private final int _minLevel;
 	private final List<ItemHolder> _itemReq;
 	private final List<Skill> _skillRem;
